+++ src/main/java/com/studleague/studleague/dao/interfaces/FlagDao.java
@@ -1,19 +1,17 @@
 package com.studleague.studleague.dao.interfaces;
 
 import com.studleague.studleague.entities.Flag;
-import com.studleague.studleague.entities.Transfer;
-
 import java.util.List;
 import java.util.Optional;
 
 public interface FlagDao {
 
-    Optional<Flag> getFlagById(int id);
+    Optional<Flag> getFlagById(long id);
 
     List<Flag> getAllFlags();
 
     void saveFlag(Flag flag);
 
-    void deleteFlag(int id);
+    void deleteFlag(long id);
 
 }
